fix a dtype bug.

+++ train/train.py
@@ -604,7 +604,7 @@ def main(args):
     vae.to(accelerator.device, dtype=weight_dtype)
     transformer.to(accelerator.device, dtype=weight_dtype)
     text_encoder_one.to(accelerator.device, dtype=weight_dtype)
-    text_encoder_two.to(accelerator.device, dtype=weight_dtype)
+    text_encoder_two.to(accelerator.device)
 
     if args.gradient_checkpointing:
         transformer.enable_gradient_checkpointing()
